CRUD implementation of Aftesia & Institucioni
As for the CRUD implementation, all necessary files were added. The user ID retrieval will be updated later. Currently, the methods are using the ID of a manually added user, as the User CRUD implementation is not complete yet. A database migration was created to change a column's data type, and a project dependency was updated. Also, mapping profiles for all entities have been created.